Add auth.types.ts
