+++ frontend/src/scripts/auth/auth.state.ts
@@ -2,22 +2,7 @@
 import { jwtDecode } from "jwt-decode";
 import { USER_URL } from "../config";
 import { logger } from "../utils/logger";
-
-/// This is the data that is stored in the JWT token to identify the user, after successful login
-export interface UserData {
-    id: string;
-    username: string;
-    displayName?: string;
-    exp: number;
-}
-
-/// This is the data that is sent to the backend when logging in or registering
-export type AuthFormData = {
-    username: string;
-    password: string;
-    displayName?: string;
-    confirmPassword?: string;
-};
+import { AuthFormData, UserData } from "./auth.types";
 
 class AuthState {
     private static instance: AuthState;
