+++ frontend/src/scripts/components/components.loginForm.ts
@@ -1,4 +1,5 @@
-import { AuthFormData, authState } from "../auth/auth.state";
+import { authState } from "../auth/auth.state";
+import { AuthFormData } from "../auth/auth.types";
 import { logger } from "../utils/logger";
 
 export const createLoginForm = async (ctaButton: HTMLElement): Promise<HTMLElement> => {
